+++ src/pharos.rs
@@ -1,4 +1,4 @@
-use crate :: { import::*, Observable, Events, ObserveConfig, events::Sender, Error, ErrorKind, Channel };
+use crate :: { import::*, Observable, Observe, Events, ObserveConfig, events::Sender, PharErr, ErrorKind, Channel };
 
 
 /// The Pharos lighthouse. When you implement [Observable] on your type, you can forward
@@ -34,17 +34,9 @@ pub struct Pharos<Event>  where Event: 'static + Clone + Send
 	//
 	observers : Vec<Option< Sender<Event> >>,
 	free_slots: Vec<usize>                  ,
-	state     : State                       ,
+	closed    : bool                        ,
 }
 
-
-#[ derive( Clone, Debug, PartialEq ) ]
-//
-enum State
-{
-	Ready,
-	Closed,
-}
 
 
 
@@ -73,7 +65,7 @@ impl<Event> Pharos<Event>  where Event: 'static + Clone + Send
 		{
 			observers : Vec::with_capacity( capacity ),
 			free_slots: Vec::with_capacity( capacity ),
-			state     : State::Ready                  ,
+			closed    : false                         ,
 		}
 	}
 
@@ -133,24 +125,24 @@ impl<Event> Default for Pharos<Event>  where Event: 'static + Clone + Send
 
 impl<Event> Observable<Event> for Pharos<Event>  where Event: 'static + Clone + Send
 {
-	type Error = Error;
+	type Error = PharErr;
 
 	/// Will re-use slots from disconnected observers to avoid growing to much.
 	///
 	/// TODO: provide API for the client to compact the pharos object after reducing the
 	///       number of observers.
 	//
-	fn observe( &mut self, options: ObserveConfig<Event> ) -> Result< Events<Event>, Self::Error >
+	fn observe( &mut self, options: ObserveConfig<Event> ) -> Observe<'_, Event, Self::Error >
 	{
-		if self.state == State::Closed
+		async move
 		{
-			return Err( ErrorKind::Closed.into() );
-		}
+			if self.closed
+			{
+				return Err( ErrorKind::Closed.into() );
+			}
 
 
-		match options.channel
-		{
-			Channel::Bounded(queue_size) =>
+			if let Channel::Bounded(queue_size) = options.channel
 			{
 				if queue_size < 1
 				{
@@ -158,26 +150,25 @@ impl<Event> Observable<Event> for Pharos<Event>  where Event: 'static + Clone + 
 				}
 			}
 
-			_ => {}
-		}
+
+			let (events, sender) = Events::new( options );
 
 
-		let (events, sender) = Events::new( options );
+			// Try to reuse a free slot
+			//
+			if let Some( i ) = self.free_slots.pop()
+			{
+				self.observers[i] = Some( sender );
+			}
 
+			else
+			{
+				self.observers.push( Some( sender ) );
+			}
 
-		// Try to reuse a free slot
-		//
-		if let Some( i ) = self.free_slots.pop()
-		{
-			self.observers[i] = Some( sender );
-		}
+			Ok( events )
 
-		else
-		{
-			self.observers.push( Some( sender ) );
-		}
-
-		Ok( events )
+		}.boxed()
 	}
 }
 
@@ -187,13 +178,13 @@ impl<Event> Observable<Event> for Pharos<Event>  where Event: 'static + Clone + 
 //
 impl<Event> Sink<Event> for Pharos<Event> where Event: Clone + 'static + Send
 {
-	type Error = Error;
+	type Error = PharErr;
 
 
 	fn poll_ready( self: Pin<&mut Self>, cx: &mut Context<'_> ) -> Poll<Result<(), Self::Error>>
 	{
 
-		if self.state == State::Closed
+		if self.closed
 		{
 			return Err( ErrorKind::Closed.into() ).into();
 		}
@@ -211,6 +202,8 @@ impl<Event> Sink<Event> for Pharos<Event> where Event: Clone + 'static + Send
 				//
 				if res.is_err()
 				{
+					// TODO: why don't we add to free_slots here like below?
+					//
 					*obs = None;
 				}
 			}
@@ -224,7 +217,7 @@ impl<Event> Sink<Event> for Pharos<Event> where Event: Clone + 'static + Send
 	fn start_send( self: Pin<&mut Self>, evt: Event ) -> Result<(), Self::Error>
 	{
 
-		if self.state == State::Closed
+		if self.closed
 		{
 			return Err( ErrorKind::Closed.into() );
 		}
@@ -271,7 +264,7 @@ impl<Event> Sink<Event> for Pharos<Event> where Event: Clone + 'static + Send
 	fn poll_flush( self: Pin<&mut Self>, cx: &mut Context<'_> ) -> Poll<Result<(), Self::Error>>
 	{
 
-		if self.state == State::Closed
+		if self.closed
 		{
 			return Err( ErrorKind::Closed.into() ).into();
 		}
@@ -308,19 +301,18 @@ impl<Event> Sink<Event> for Pharos<Event> where Event: Clone + 'static + Send
 
 
 
-	/// Will close and drop all observers. The pharos object will remain operational however.
-	/// The main annoyance would be that we'd have to make
+	/// Will close and drop all observers.
 	//
 	fn poll_close( mut self: Pin<&mut Self>, cx: &mut Context<'_> ) -> Poll<Result<(), Self::Error>>
 	{
-		if self.state == State::Closed
+		if self.closed
 		{
 			return Ok(()).into();
 		}
 
 		else
 		{
-			self.state = State::Closed;
+			self.closed = true;
 		}
 
 
@@ -401,9 +393,9 @@ mod tests
 
 	// verify storage_len and num_observers
 	//
-	#[test]
+	#[async_std::test]
 	//
-	fn storage_len()
+	async fn storage_len()
 	{
 		let mut ph = Pharos::<bool>::default();
 
@@ -411,13 +403,13 @@ mod tests
 			assert_eq!( ph.num_observers (), 0 );
 			assert_eq!( ph.free_slots.len(), 0 );
 
-		let mut a = ph.observe( ObserveConfig::default() ).expect( "observe" );
+		let mut a = ph.observe( ObserveConfig::default() ).await.expect( "observe" );
 
 			assert_eq!( ph.storage_len   (), 1 );
 			assert_eq!( ph.num_observers (), 1 );
 			assert_eq!( ph.free_slots.len(), 0 );
 
-		let b = ph.observe( ObserveConfig::default() ).expect( "observe" );
+		let b = ph.observe( ObserveConfig::default() ).await.expect( "observe" );
 
 			assert_eq!( ph.storage_len   (), 2 );
 			assert_eq!( ph.num_observers (), 2 );
@@ -439,14 +431,14 @@ mod tests
 
 	// observe: Make sure we are reusing slots
 	//
-	#[test]
+	#[async_std::test]
 	//
-	fn reuse()
+	async fn reuse()
 	{
 		let mut ph = Pharos::<bool>::default();
-		let _a = ph.observe( ObserveConfig::default() );
-		let  b = ph.observe( ObserveConfig::default() );
-		let _c = ph.observe( ObserveConfig::default() );
+		let _a = ph.observe( ObserveConfig::default() ).await;
+		let  b = ph.observe( ObserveConfig::default() ).await;
+		let _c = ph.observe( ObserveConfig::default() ).await;
 
 			assert_eq!( ph.storage_len  (), 3 );
 			assert_eq!( ph.num_observers(), 3 );
@@ -462,13 +454,13 @@ mod tests
 			assert_eq!( &ph.free_slots, &[1] );
 
 
-		let _d = ph.observe( ObserveConfig::default() );
+		let _d = ph.observe( ObserveConfig::default() ).await;
 
 			assert_eq!( ph.storage_len   (), 3 );
 			assert_eq!( ph.num_observers (), 3 );
 			assert_eq!( ph.free_slots.len(), 0 );
 
-		let _e = ph.observe( ObserveConfig::default() );
+		let _e = ph.observe( ObserveConfig::default() ).await;
 
 			// Now we should have pushed again
 			//
@@ -480,90 +472,94 @@ mod tests
 
 	// observe: verify we can no longer observe after calling close
 	//
-	#[test]
+	#[async_std::test]
 	//
-	fn observe_after_close()
+	async fn observe_after_close()
 	{
 		let mut ph = Pharos::<bool>::default();
 
 		block_on( ph.close() ).expect( "close" );
 
-		let res = ph.observe( ObserveConfig::default() );
+		let res = ph.observe( ObserveConfig::default() ).await;
 
-		assert!   ( res.is_err() );
-		assert_eq!( ErrorKind::Closed, res.unwrap_err().kind() );
+			assert!   ( res.is_err() );
+			assert_eq!( ErrorKind::Closed, res.unwrap_err().kind() );
 	}
 
 
 	// observe: refuse Channel::Bounded(0)
 	//
-	#[test]
+	#[async_std::test]
 	//
-	fn observe_refuse_zero()
+	async fn observe_refuse_zero()
 	{
 		let mut ph = Pharos::<bool>::default();
 
-		let res = ph.observe( Channel::Bounded(0).into() );
+		let res = ph.observe( Channel::Bounded(0).into() ).await;
 
-		assert!   ( res.is_err() );
-		assert_eq!( ErrorKind::MinChannelSizeOne, res.unwrap_err().kind() );
+			assert!   ( res.is_err() );
+			assert_eq!( ErrorKind::MinChannelSizeOne, res.unwrap_err().kind() );
 	}
 
 
 	// verify that one observer blocks pharos.
 	//
-	#[ test ]
+	#[async_std::test]
 	//
-	fn poll_ready_pending()
+	async fn poll_ready_pending()
 	{
-		block_on( poll_fn( move |mut cx|
+		let mut ph = Pharos::default();
+
+		let _open    = ph.observe( Channel::Bounded  ( 10 ).into() ).await.expect( "observe" );
+		let mut full = ph.observe( Channel::Bounded  ( 1  ).into() ).await.expect( "observe" );
+		let _unbound = ph.observe( Channel::Unbounded      .into() ).await.expect( "observe" );
+
+		poll_fn( move |mut cx|
 		{
-			let mut ph = Pharos::default();
-
-			let _open    = ph.observe( Channel::Bounded  ( 10 ).into() ).expect( "observe" );
-			let mut full = ph.observe( Channel::Bounded  ( 1  ).into() ).expect( "observe" );
-			let _unbound = ph.observe( Channel::Unbounded      .into() ).expect( "observe" );
-
 			let mut ph = Pin::new( &mut ph );
 
-			assert_matches!( ph.as_mut().poll_ready( &mut cx ), Poll::Ready( Ok(_) ) );
-			assert!( ph.as_mut().start_send( true ).is_ok() );
+				assert_matches!( ph.as_mut().poll_ready( &mut cx ), Poll::Ready( Ok(_) ) );
+				assert!( ph.as_mut().start_send( true ).is_ok() );
 
-			assert_matches!( ph.as_mut().poll_ready( &mut cx ), Poll::Pending );
+				assert_matches!( ph.as_mut().poll_ready( &mut cx ), Poll::Pending );
 
-			assert_eq!( Pin::new( &mut full ).poll_next(cx), Poll::Ready(Some(true)));
+				assert_eq!( Pin::new( &mut full ).poll_next(cx), Poll::Ready(Some(true)));
 
-			assert_matches!( ph.as_mut().poll_ready( &mut cx ), Poll::Ready( Ok(_) ) );
+				assert_matches!( ph.as_mut().poll_ready( &mut cx ), Poll::Ready( Ok(_) ) );
 
 			().into()
-		}));
+
+		}).await;
 	}
 
 
 
 	// pharos drops closed observers.
 	//
-	#[ test ]
+	#[async_std::test]
 	//
-	fn poll_ready_drop()
+	async fn poll_ready_drop()
 	{
-		block_on( poll_fn( move |mut cx|
+		let mut ph = Pharos::<bool>::default();
+
+		let _open    = ph.observe( Channel::Bounded  ( 10 ).into() ).await.expect( "observe" );
+		let full     = ph.observe( Channel::Bounded  ( 1  ).into() ).await.expect( "observe" );
+		let _unbound = ph.observe( Channel::Unbounded      .into() ).await.expect( "observe" );
+
+		let mut ph = Pin::new( &mut ph );
+
+		drop( full );
+
+
+		poll_fn( move |mut cx|
 		{
-			let mut ph = Pharos::<bool>::default();
-
-			let _open    = ph.observe( Channel::Bounded  ( 10 ).into() ).expect( "observe" );
-			let full     = ph.observe( Channel::Bounded  ( 1  ).into() ).expect( "observe" );
-			let _unbound = ph.observe( Channel::Unbounded      .into() ).expect( "observe" );
-
-			let mut ph = Pin::new( &mut ph );
-
-			drop( full );
-
 			assert_matches!( ph.as_mut().poll_ready( &mut cx ), Poll::Ready( Ok(_) ) );
 
 			assert!( ph.observers[1].is_none() );
+
 			().into()
-		}));
+
+		}).await;
 	}
 
 
@@ -588,8 +584,8 @@ mod tests
 
 				match res
 				{
-					Poll::Ready( Err( e ) ) => assert_eq!( ErrorKind::Closed, e.kind() ),
-					_                       => assert!( false, "wrong result " ),
+					Poll::Ready( Err( e ) ) => assert_eq!( ErrorKind::Closed, e.kind() ) ,
+					_                       => unreachable!( "wrong result " )           ,
 				}
 
 			().into()
@@ -601,18 +597,18 @@ mod tests
 
 	// start_send verify message arrives.
 	//
-	#[ test ]
+	#[async_std::test]
 	//
-	fn start_send_arrive()
+	async fn start_send_arrive()
 	{
-		block_on( poll_fn( move |mut cx|
+		let mut ph = Pharos::<usize>::default();
+
+		let _open    = ph.observe( Channel::Bounded  ( 10 ).into() ).await.expect( "observe" );
+		let mut full = ph.observe( Channel::Bounded  ( 1  ).into() ).await.expect( "observe" );
+		let _unbound = ph.observe( Channel::Unbounded      .into() ).await.expect( "observe" );
+
+		poll_fn( move |mut cx|
 		{
-			let mut ph = Pharos::default();
-
-			let _open    = ph.observe( Channel::Bounded  ( 10 ).into() ).expect( "observe" );
-			let mut full = ph.observe( Channel::Bounded  ( 1  ).into() ).expect( "observe" );
-			let _unbound = ph.observe( Channel::Unbounded      .into() ).expect( "observe" );
-
 			let mut ph = Pin::new( &mut ph );
 
 			assert_matches!( ph.as_mut().poll_ready( &mut cx ), Poll::Ready( Ok(_) ) );
@@ -621,34 +617,36 @@ mod tests
 			assert_eq!( Pin::new( &mut full ).poll_next(cx), Poll::Ready(Some(3)));
 
 			().into()
-		}));
+
+		}).await;
 	}
 
 
 
 	// pharos drops closed observers.
 	//
-	#[ test ]
+	#[async_std::test]
 	//
-	fn poll_flush_drop()
+	async fn poll_flush_drop()
 	{
-		block_on( poll_fn( move |mut cx|
+		let mut ph = Pharos::<bool>::default();
+
+		let _open    = ph.observe( Channel::Bounded  ( 10 ).into() ).await.expect( "observe" );
+		let full     = ph.observe( Channel::Bounded  ( 1  ).into() ).await.expect( "observe" );
+		let _unbound = ph.observe( Channel::Unbounded      .into() ).await.expect( "observe" );
+
+		let mut ph = Pin::new( &mut ph );
+
+		drop( full );
+
+		poll_fn( move |mut cx|
 		{
-			let mut ph = Pharos::<bool>::default();
-
-			let _open    = ph.observe( Channel::Bounded  ( 10 ).into() ).expect( "observe" );
-			let full     = ph.observe( Channel::Bounded  ( 1  ).into() ).expect( "observe" );
-			let _unbound = ph.observe( Channel::Unbounded      .into() ).expect( "observe" );
-
-			let mut ph = Pin::new( &mut ph );
-
-			drop( full );
-
 			assert_matches!( ph.as_mut().poll_flush( &mut cx ), Poll::Ready( Ok(_) ) );
 
 			assert!( ph.observers[1].is_none() );
 			().into()
-		}));
+
+		}).await;
 	}